+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 * Removed d3-tip from the dependencies.
 * Added AMD and CommonJS support.
 * Fixed a bug that was causing the resize of the SVG to calculate the width and height incorrectly.
+* Made *data* return the RelationshipGraph object to keep d3's chaining functionality working.